edd/requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
tzdata>=2024.1
tiktoken
